@@ -18,46 +18,43 @@
   "nativescript": {
     "id": "org.nativescript.groceries",
     "tns-android": {
-      "version": "3.0.1"
+      "version": "3.2.0"
     },
     "tns-ios": {
-      "version": "3.0.1"
+      "version": "3.2.0"
     }
   },
   "scripts": {
     "tslint": "tslint -p tsconfig.json",
     "ns-bundle": "ns-bundle",
     "publish-ios-bundle": "npm run ns-bundle --ios --publish-app",
+    "generate-android-snapshot": "generate-android-snapshot --targetArchs arm,arm64,ia32 --install",
     "start-android-bundle": "npm run ns-bundle --android --run-app",
     "start-ios-bundle": "npm run ns-bundle --ios --run-app",
     "build-android-bundle": "npm run ns-bundle --android --build-app",
-    "build-ios-bundle": "npm run ns-bundle --ios --build-app",
-    "generate-android-snapshot": "generate-android-snapshot --projectRoot . --targetArchs arm,arm64 --install"
+    "build-ios-bundle": "npm run ns-bundle --ios --build-app"
   },
   "dependencies": {
-    "@angular/common": "~4.1.3",
-    "@angular/compiler": "~4.1.3",
-    "@angular/core": "~4.1.3",
-    "@angular/forms": "~4.1.3",
-    "@angular/http": "~4.1.3",
-    "@angular/platform-browser": "~4.1.3",
-    "@angular/router": "~4.1.3",
+    "@angular/common": "~4.4.5",
+    "@angular/compiler": "~4.4.5",
+    "@angular/core": "~4.4.5",
+    "@angular/forms": "~4.4.5",
+    "@angular/http": "~4.4.5",
+    "@angular/platform-browser": "~4.4.5",
+    "@angular/router": "~4.4.5",
     "email-validator": "^1.0.7",
-    "kinvey-nativescript-sdk": "3.7.0",
-    "nativescript-angular": "next",
-    "nativescript-iqkeyboardmanager": "1.0.1",
-    "nativescript-pulltorefresh": "2.0.1",
-    "nativescript-social-share": "~1.3.2",
-    "nativescript-sqlite": "1.1.7",
+    "nativescript-angular": "~4.4.1",
+    "nativescript-iqkeyboardmanager": "~1.1.0",
+    "nativescript-social-share": "~1.4.0",
     "nativescript-unit-test-runner": "^0.3.3",
-    "reflect-metadata": "^0.1.8",
-    "rxjs": "^5.4.0",
-    "tns-core-modules": "next",
-    "zone.js": "^0.8.11"
+    "reflect-metadata": "~0.1.8",
+    "rxjs": "~5.4.2",
+    "tns-core-modules": "~3.2.0",
+    "zone.js": "~0.8.2"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "~4.1.3",
-    "@ngtools/webpack": "~1.4.0",
+    "@angular/compiler-cli": "~4.4.5",
+    "@ngtools/webpack": "~1.5.3",
     "babel-traverse": "6.24.1",
     "babel-types": "6.24.1",
     "babylon": "6.17.0",
@@ -71,13 +68,15 @@
     "karma-nativescript-launcher": "^0.4.0",
     "lazy": "1.0.11",
     "nativescript-css-loader": "~0.26.0",
-    "nativescript-dev-typescript": "^0.4.5",
-    "nativescript-dev-webpack": "next",
+    "nativescript-dev-typescript": "~0.5.0",
+    "nativescript-dev-webpack": "~0.8.0",
     "raw-loader": "~0.5.1",
     "resolve-url-loader": "~2.0.2",
     "tslint": "^5.4.2",
-    "typescript": "~2.3.4",
+    "typescript": "~2.4.2",
     "webpack": "~2.6.1",
-    "webpack-sources": "~1.0.1"
+    "webpack-bundle-analyzer": "^2.8.2",
+    "webpack-sources": "~1.0.1",
+    "nativescript-worker-loader": "~0.8.1"
   }
 }
