@@ -5,6 +5,7 @@ const nsWebpack = require("nativescript-dev-webpack");
 const nativescriptTarget = require("nativescript-dev-webpack/nativescript-target");
 const CopyWebpackPlugin = require("copy-webpack-plugin");
 const ExtractTextPlugin = require("extract-text-webpack-plugin");
+const { BundleAnalyzerPlugin } = require("webpack-bundle-analyzer");
 
 const { AotPlugin } = require("@ngtools/webpack");
 
@@ -48,7 +49,11 @@ module.exports = env => {
             modules: [
                 "node_modules/tns-core-modules",
                 "node_modules",
-            ]
+            ],
+
+            alias: {
+                '~': resolve("./app")
+            },
         },
         node: {
             // Disable node shims that conflict with NativeScript
@@ -66,7 +71,7 @@ module.exports = env => {
             chunk: "vendor",
             projectRoot: __dirname,
             webpackConfig: config,
-            targetArchs: ["arm", "arm64"],
+            targetArchs: ["arm", "arm64", "ia32"],
             tnsJavaClassesOptions: { packages: ["tns-core-modules" ] },
             useLibs: false
         }));
@@ -166,6 +171,15 @@ function getPlugins(platform, env) {
             "./bundle",
         ]),
 
+        // Generate report files for bundles content
+        new BundleAnalyzerPlugin({
+            analyzerMode: "static",
+            openAnalyzer: false,
+            generateStatsFile: true,
+            reportFilename: join(__dirname, "report", `${platform}-report.html`),
+            statsFilename: join(__dirname, "report", `${platform}-stats.json`),
+        }),
+
         // Angular AOT compiler
         new AotPlugin({
             tsConfigPath: "tsconfig.aot.json",
@@ -173,11 +187,15 @@ function getPlugins(platform, env) {
             typeChecking: false
         }),
 
-        // Resolve .ios.css and .android.css component stylesheets
-        new nsWebpack.UrlResolvePlugin({platform}),
+        // Resolve .ios.css and .android.css component stylesheets, and .ios.html and .android component views
+        new nsWebpack.UrlResolvePlugin({
+            platform: platform,
+            resolveStylesUrls: true,
+            resolveTemplateUrl: true
+        }),
 
     ];
-    
+
     if (env.uglify) {
         plugins.push(new webpack.LoaderOptionsPlugin({ minimize: true }));
 
